Changed functions to work with long long ints not only ints. Ft_Printf not changed yet due to long logic problems, so when printing the list using my ft_printf i cant use %lld to print the long long int and it rolls back if the num is over maxint

+++ include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: slegaris <slegaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/12 17:30:33 by slegaris          #+#    #+#             */
-/*   Updated: 2023/10/17 22:26:56 by slegaris         ###   ########.fr       */
+/*   Updated: 2023/10/17 23:23:50 by slegaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,12 +21,12 @@
 
 typedef struct s_stack
 {
-	int content;
+	long long int	content;
 	int index;
 	struct s_stack *next;
 } t_stack;
 
-t_stack	*ft_newstack(int content);
+t_stack	*ft_newstack(long long int content);
 void	ft_lstaddback(t_stack **lst, t_stack *newer);
 t_stack	*ft_lst_last(t_stack *lst);
 void	ft_printlist(t_stack **list);
