+++ src/index.c
@@ -6,7 +6,7 @@
 /*   By: slegaris <slegaris@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/17 20:30:33 by slegaris          #+#    #+#             */
-/*   Updated: 2023/10/17 22:30:29 by slegaris         ###   ########.fr       */
+/*   Updated: 2023/10/17 23:18:31 by slegaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@ void	lstindex(t_stack *lst)
 	t_stack	*mid;
 	int		i;
 	int		nodes;
-	int		mini;
+	long	mini;
 
 	i = 0;
 	nodes = ft_stacksize(lst);
 	while (i <= nodes)
 	{
 		current = lst;
-		mini = INT_MAX;
+		mini = LONG_MAX;
 		mid = NULL;
 		while (current)
 		{
